api/pyproject.toml
[project]
name = "api"
version = "0.1.1"
description = ""
authors = [{ name = "levch", email = "levchenko.d.a1998@gmail.com" }]
dependencies = [
    "i-texts>=0.1.1",
    "mongoengine>=0.27.0",
    "cloudipsp>=1.0.4",
    "parsenv>=0.1.3",
    "aiohttp>=3.8.4",
    "pillow>=9.5.0",
    "pydantic>=1.10.7",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts._]
env_file = "../.env"
env = { PYTHONPATH = "." }

[tool.pdm.scripts]
test = "python tests"

[tool.ruff]
select = ["ALL"]
ignore = ["D", "ANN", "F401", "DTZ"]
fix = true
src = ["api", "tests"]

[tool.pyright]
#typeCheckingMode = "strict"

[tool.isort]
profile = "black"